social_media/requirements.txt
atproto
beautifulsoup4
python-dotenv
requests
pytumblr
